Move file out of scripts per code review suggestion
